backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==6.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==21.2.0
openpyxl==3.1.2
pandas==2.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.6.0
